Moving subtopic methods to subtopicController
