+++ hydra-topic-service/src/main/java/com/revature/beans/Topic.java
@@ -1,4 +1,4 @@
-package com.revature.model;
+package com.revature.beans;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
