+++ hydra-topic-service/src/main/java/com/revature/repository/TopicRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.revature.model.Topic;
+import com.revature.beans.Topic;
 
 
 @Repository
