+++ hydra-topic-service/src/main/java/com/revature/services/TopicService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.revature.model.Topic;
+import com.revature.beans.Topic;
 import com.revature.repository.TopicRepository;
 
 @Service
